+++ conseil-api/src/main/scala/tech/cryptonomic/conseil/api/ConseilApi.scala
@@ -4,24 +4,23 @@ import java.util.UUID
 
 import akka.actor.ActorSystem
 import akka.event.Logging
+import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import akka.stream.ActorMaterializer
 import cats.effect.{ContextShift, IO}
 import ch.megard.akka.http.cors.scaladsl.CorsDirectives.cors
-import akka.http.scaladsl.server.Directives._
 import tech.cryptonomic.conseil.api.config.ConseilAppConfig.CombinedConfiguration
 import tech.cryptonomic.conseil.api.directives.{EnableCORSDirectives, RecordingDirectives, ValidatingDirectives}
 import tech.cryptonomic.conseil.api.metadata.{AttributeValuesCacheConfiguration, MetadataService, UnitTransformation}
 import tech.cryptonomic.conseil.api.routes.Docs
 import tech.cryptonomic.conseil.api.routes.info.AppInfo
-import tech.cryptonomic.conseil.api.routes.platform.data.ApiDataEndpoints
+import tech.cryptonomic.conseil.api.routes.platform.data.ApiDataRoutes
 import tech.cryptonomic.conseil.api.routes.platform.discovery.PlatformDiscovery
 import tech.cryptonomic.conseil.api.routes.platform.{Api, TezosApi}
 import tech.cryptonomic.conseil.api.security.Security
 import tech.cryptonomic.conseil.common.config.Platforms
 import tech.cryptonomic.conseil.common.config.Platforms.BlockchainPlatform
 import tech.cryptonomic.conseil.common.generic.chain.PlatformDiscoveryOperations
-import tech.cryptonomic.conseil.common.generic.chain.PlatformDiscoveryTypes.Platform
 
 import scala.concurrent.ExecutionContext
 
@@ -34,8 +33,6 @@ class ConseilApi(config: CombinedConfiguration)(implicit system: ActorSystem) ex
   private val transformation = new UnitTransformation(config.metadata)
   private val cacheOverrides = new AttributeValuesCacheConfiguration(config.metadata)
 
-  private val apiCache = new ApiCache()
-
   implicit private val mat: ActorMaterializer = ActorMaterializer()
   implicit private val dispatcher: ExecutionContext = system.dispatcher
   implicit private val contextShift: ContextShift[IO] = IO.contextShift(dispatcher)
@@ -46,7 +43,7 @@ class ConseilApi(config: CombinedConfiguration)(implicit system: ActorSystem) ex
 
   // this val is not lazy to force to fetch metadata and trigger logging at the start of the application
   private val metadataService =
-    new MetadataService(config.platforms, transformation, cacheOverrides, apiCache.cachedDiscoveryOperations)
+    new MetadataService(config.platforms, transformation, cacheOverrides, ApiCache.cachedDiscoveryOperations)
 
   lazy val route: Route = {
     lazy val platformDiscovery = PlatformDiscovery(metadataService)
@@ -84,7 +81,7 @@ class ConseilApi(config: CombinedConfiguration)(implicit system: ActorSystem) ex
                           logRequest("Metadata Route", Logging.DebugLevel) {
                             platformDiscovery.route
                           },
-                          apiCache
+                          ApiCache
                             .cachedDataEndpoints(metadataService)
                             .map {
                               case (platform, routes) =>
@@ -109,16 +106,30 @@ class ConseilApi(config: CombinedConfiguration)(implicit system: ActorSystem) ex
     )
   }
 
-  private class ApiCache {
+  /**
+    * Object, which initializes and holds all of the APIs (blockchain-specific endpoints) in the map.
+    *
+    * Note that only APIs that are visible (configured via metadata config file)
+    * will be initialized and eventually exposed.
+    */
+  private object ApiCache {
     private lazy val cache: Map[BlockchainPlatform, Api] = forVisiblePlatforms {
       case Platforms.Tezos => new TezosApi(config.metadata, config.server)
     }
 
+    /**
+      * Map, that contains list of available `PlatformDiscoveryOperations` accessed by platform name.
+      * @see `tech.cryptonomic.conseil.common.config.Platforms` to get list of possible platforms.
+      */
     lazy val cachedDiscoveryOperations: Map[String, PlatformDiscoveryOperations] = cache.map {
       case (key, value) => key.name -> value.discoveryOperations
     }
 
-    lazy val cachedDataEndpoints: MetadataService => Map[String, ApiDataEndpoints] = service =>
+    /**
+      * Map, that contains list of available `ApiDataRoutes` accessed by platform name.
+      * @see `tech.cryptonomic.conseil.common.config.Platforms` to get list of possible platforms.
+      */
+    lazy val cachedDataEndpoints: MetadataService => Map[String, ApiDataRoutes] = service =>
       cache.map {
         case (key, value) => key.name -> value.dataEndpoint(service)
       }
@@ -126,11 +137,11 @@ class ConseilApi(config: CombinedConfiguration)(implicit system: ActorSystem) ex
     private def forVisiblePlatforms(f: BlockchainPlatform => Api): Map[BlockchainPlatform, Api] =
       transformation
         .overridePlatforms(config.platforms.getPlatforms)
-        .map(toBlockchainPlatform)
-        .map(platform => platform -> f(platform))
+        .map { platform =>
+          val blockchainPlatform = BlockchainPlatform.fromString(platform.name)
+          blockchainPlatform -> f(blockchainPlatform)
+        }
         .toMap
-
-    private def toBlockchainPlatform(v: Platform): BlockchainPlatform = BlockchainPlatform.fromString(v.name)
   }
 
 }
